add manage codes container
